+++ index.html
@@ -25,6 +25,11 @@
           options="이름순,거리순"
         ></filter-list>
       </section>
+      <section class="restaurant-list-container">
+        <ul class="restaurant-list">
+          <restaurant-box></restaurant-box>
+        </ul>
+      </section>
       <restaurant-modal></restaurant-modal>
     </div>
   </body>
